Change `get_interior_ring_n` parameter dtype to usize

+++ src/geometry.rs
@@ -1392,7 +1392,7 @@ pub trait Geom: AsRaw<RawType = GEOSGeometry> {
     ///                         1.0000000000000000 5.0000000000000000, \
     ///                         1.0000000000000000 1.0000000000000000)");
     /// ```
-    fn get_interior_ring_n(&self, n: u32) -> GResult<ConstGeometry>;
+    fn get_interior_ring_n(&self, n: usize) -> GResult<ConstGeometry>;
     /// Returns the exterior ring.
     ///
     /// # Example
@@ -2339,7 +2339,7 @@ impl$(<$lt>)? Geom for $ty_name$(<$lt>)? {
         })
     }
 
-    fn get_interior_ring_n(&self, n: u32) -> GResult<ConstGeometry> {
+    fn get_interior_ring_n(&self, n: usize) -> GResult<ConstGeometry> {
         with_context(|ctx| unsafe {
             let ptr = GEOSGetInteriorRingN_r(ctx.as_raw(), self.as_raw(), n as _);
             ConstGeometry::new_from_raw(ptr, ctx, self$(.$field)?, "get_interior_ring_n")
